Fix some typos + more coversions to CGFloat from Float

+++ LazyTransitions/PartialTransitionAnimator.swift
@@ -11,10 +11,10 @@ import UIKit
 public class PartialTransitionAnimator: NSObject {
     public weak var delegate : TransitionAnimatorDelegate?
     public var orientation : TransitionOrientation
-    public var speed : Float
+    public var speed : CGFloat
     public var allowedOrientations: [TransitionOrientation]?
     public var supportedOrientations: [TransitionOrientation] = [.topToBottom, .bottomToTop, .leftToRight, .rightToLeft]
-    public init(orientation: TransitionOrientation, speed: Float) {
+    public init(orientation: TransitionOrientation, speed: CGFloat) {
         self.orientation = orientation
         self.speed = speed
     }
@@ -39,7 +39,7 @@ extension PartialTransitionAnimator: TransitionAnimator {
         let containerView = transitionContext.containerView
         containerView.insertSubview(toVC.view, belowSubview: fromVC.view)
         
-        var adjustmentRatio = CGFloat(self.speed / 25)
+        var adjustmentRatio = self.speed / 25
         
         if adjustmentRatio > 0.3 {
             adjustmentRatio = 0.3
